+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Fixed
+
+- Fixed bug where closing the window on macOS and then re-opening would break the IPC between main and renderer processes.
+
 ## [5.8.1] - 2025-10-16
 
 ### Fixed