+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## [5.8.2] - 2025-10-20
+
 ### Fixed
 
 - Fixed bug where closing the window on macOS and then re-opening would break the IPC between main and renderer processes.
@@ -968,7 +970,9 @@ Fixed bug where pressing Ctrl-Shift-C to copy text from a terminal would enable 
 - Added auto-scroll to TX pane, closes #89.
 - Added special delete behaviour for backspace button when in "send on enter" mode, closes #90.
 
-[unreleased]: https://github.com/gbmhunter/NinjaTerm/compare/v5.8.0...HEAD
+[unreleased]: https://github.com/gbmhunter/NinjaTerm/compare/v5.8.2...HEAD
+[5.8.2]: https://github.com/gbmhunter/NinjaTerm/compare/v5.8.1...v5.8.2
+[5.8.1]: https://github.com/gbmhunter/NinjaTerm/compare/v5.8.0...v5.8.1
 [5.8.0]: https://github.com/gbmhunter/NinjaTerm/compare/v5.7.1...v5.8.0
 [5.7.1]: https://github.com/gbmhunter/NinjaTerm/compare/v5.7.0...v5.7.1
 [5.7.0]: https://github.com/gbmhunter/NinjaTerm/compare/v5.6.0...v5.7.0